requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-Bcrypt
Flask-JWT-Extended
python-dotenv
